avancement

+++ javaCode/LogicielGestion/src/logicielgestion/Commercial.java
@@ -13,7 +13,7 @@ public class Commercial extends Employes {
     //volume mensuel
     private final int volume;
     //CONSTRUCTEUR 
-    public Commercial(String n, String p, int mat,int indice, int v) {
+    public Commercial(String n, String p, int mat,float indice, int v) {
         super(n, p, mat,indice);
         this.volume=v;
     }
